+++ kiosk/kiosk_browser/browser_widget.py
@@ -2,6 +2,7 @@ from PyQt5 import QtCore, QtWidgets, QtWebEngineWidgets, QtGui, QtSvg
 from enum import Enum, auto
 import logging
 import re
+import time
 
 from kiosk_browser import system
 
@@ -59,8 +60,11 @@ class BrowserWidget(QtWidgets.QWidget):
         self._webview.loadFinished.connect(self._load_finished)
 
         # Shortcut to manually reload
-        self.reload_shortcut = QtWidgets.QShortcut('CTRL+R', self)
-        self.reload_shortcut.activated.connect(self.reload)
+        self._reload_shortcut = QtWidgets.QShortcut('CTRL+R', self)
+        self._reload_shortcut.activated.connect(self.reload)
+        # Shortcut to perform a hard refresh
+        self._hard_refresh_shortcut = QtWidgets.QShortcut('CTRL+SHIFT+R', self)
+        self._hard_refresh_shortcut.activated.connect(self._hard_refresh)
 
         # Prepare reload timer
         self._reload_timer = QtCore.QTimer(self)
@@ -92,6 +96,23 @@ class BrowserWidget(QtWidgets.QWidget):
         if not success:
             self._view(Status.NETWORK_ERROR)
             self._reload_timer.start(reload_on_network_error_after)
+
+    def _hard_refresh(self):
+        """ Clear cache, then reload.
+
+        Does not affect cookies or localstorage contents.
+
+        NOTE This clears the entire HTTP cache, assumed to be OK as the kiosk targets a specific page.
+        """
+        logging.info(f"Clearing HTTP cache (hard refresh)")
+        self._webview.page().profile().clearHttpCache()
+
+        # Sleep before triggering reload to avoid a possible race condition.
+        # Future Qt versions may provide a signal on `QWebEngineProfile` to
+        # allow to queue the reload on clear-cache completion instead.
+        # https://bugreports.qt.io/browse/QTBUG-111541
+        time.sleep(0.25)
+        self.reload()
 
     def _proxy_auth(self, get_current_proxy, url, auth, proxyHost):
         proxy = get_current_proxy()